refact: type declaration extracted from each files
