+++ src/components/ThemePanels/ThemeContents.tsx
@@ -1,18 +1,9 @@
-import { useContext, useEffect } from "react";
-import { useNavigate, useParams } from "react-router-dom";
+import { useParams } from "react-router-dom";
 import { fetchThemeInfo } from "@src/apis/BackEnd/apiList";
 import HeroPanel from "@src/components/ThemePanels/HeroPanel";
-import ToastContext from "@src/contexts/ToastContext";
 import { useSuspenseQuery } from "@tanstack/react-query";
 import ThemePanel from "./ThemePanel";
-
-export type ThemeInfo = {
-  backgroundColor: string;
-  description: number;
-  name: string;
-  themeId: string;
-  title: number;
-};
+import type { ThemeInfo } from "@src/types/ThemeTypes";
 
 function ThemeContents() {
   const themeId = useParams().id ?? "";
