+++ src/contexts/UserContext.tsx
@@ -1,11 +1,5 @@
-import type { StateHook } from "@src/hooks/stateHookType";
+import type { UserContextType } from "@src/types/UserContextType";
 import { createContext } from "react";
-
-type UserContextType = {
-  authToken: StateHook<string | null>;
-  user: StateHook<string | null>;
-  email: StateHook<string | null>;
-} | null;
 
 const UserContext = createContext<UserContextType>(null);
