+++ src/types/OrderFormTypes.ts
@@ -1,0 +1,13 @@
+export type Receiver = {
+  id: string;
+  name: string;
+  phoneNumber: string;
+  quantity: string;
+};
+
+export type FormType = {
+  cardId: string;
+  message: string;
+  sender: string;
+  receivers: Receiver[];
+};
