+++ src/types/ProductDataType.ts
@@ -1,0 +1,7 @@
+export type ProductData = {
+    imageURL: string;
+    id: number;
+    name: string;
+    brandName: string;
+    price: number;
+  };
