+++ src/types/ProductInfoType.ts
