+++ src/types/ToastContextType.ts
@@ -1,0 +1,5 @@
+import type { StateHook } from "@src/hooks/stateHookType";
+
+export type ToastContextType = {
+  message: StateHook<string | null>;
+} | null;
