+++ src/types/UserContextType.ts
@@ -1,0 +1,7 @@
+import type { StateHook } from "@src/hooks/stateHookType";
+
+export type UserContextType = {
+  authToken: StateHook<string | null>;
+  user: StateHook<string | null>;
+  email: StateHook<string | null>;
+} | null;
